refactor: extract BlockObserverConfig
