+++ libs/jwst/src/workspaces/mod.rs
@@ -4,6 +4,7 @@ mod plugins;
 mod sync;
 mod transaction;
 mod workspace;
+mod block_observer;
 
 use super::{constants, error, info, trace, warn, JwstError, JwstResult, Space};
 
@@ -11,4 +12,5 @@ pub use metadata::{Pages, WorkspaceMetadata};
 #[cfg(feature = "workspace-search")]
 pub use plugins::{SearchResult, SearchResults};
 pub use transaction::WorkspaceTransaction;
-pub use workspace::{MapSubscription, Workspace, BlockObserverConfig};
+pub use workspace::{MapSubscription, Workspace};
+pub use block_observer::BlockObserverConfig;
